v1.3-Pre3
This commit changes and adds quite a few things.

Additions:
+ The client will now display a notification if it fails to load or save a hotbar file properly.
+ Adds Javadoc documentation for many parts of the mod's code.

Changes:
* Increases the hotbar page capacity from 2,147,483,648 total pages to a whopping 9,223,372,036,854,775,808 total pages.
* Updates the README document.
* Upticks the mod version from v1.3-Pre2 to v1.3-Pre3.
* General code cleanup

Removals:
- Removes More Toolbars from the list of incompatible mods. While it is still not recommended that you use the two together, it can be done if you *really* need the extra capacity.
- Removes the repository used for LibGui (again).

+++ src/main/java/me/videogamesm12/multihotbar/commands/PageCommand.java
@@ -37,7 +37,7 @@ public class PageCommand implements Command<FabricClientCommandSource>
 	@Override
 	public int run(CommandContext<FabricClientCommandSource> context) throws CommandSyntaxException
 	{
-		Integer page = context.getArgument("page", Integer.class);
+		Long page = context.getArgument("page", Long.class);
 
 		// Prevents the player from going into the negative pages.
 		// TODO: Make this a bit cleaner by using command exceptions instead. Consider removing this check in the future
